issue #3 - idx 수정

+++ src/main/java/com/example/animore/Search/LocationController.java
@@ -20,7 +20,7 @@ public class LocationController {
     }
 
     @GetMapping("/{locationId}")
-    public Location getLocation(@PathVariable int locationId) {
+    public Location getLocation(@PathVariable Long locationId) {
         return locationRepository.findByLocationId(locationId);
     }
 }
