+++ src/main/java/com/example/animore/Search/model/SearchHistory.java
@@ -20,9 +20,9 @@ public class SearchHistory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "search_id")
-    private int searchId;
+    private Long searchId;
     @Column(name = "user_idx")
-    private int userIdx;
+    private Long userIdx;
     @Column(name = "search_query")
     private String searchQuery;
     @Column(name = "search_create_at")
